refactor(services): R-06 — extract reverseProxy.ts helper for embed route reuse
Move all proxy logic from the route handler into src/lib/services/reverseProxy.ts,
exposing a generic proxyRequest() + ReverseProxyConfig interface. The route handler
becomes a thin wrapper (~60 LOC). All 20 embed-proxy tests pass unchanged.
