+++ src/app/(dashboard)/dashboard/providers/services/[name]/embed/[...path]/route.ts
@@ -3,203 +3,76 @@
  *
  * Route: /dashboard/providers/services/[name]/embed/[...path]
  *
- * Forwards HTTP traffic to a locally-running embedded service (e.g. 9router
- * listening on 127.0.0.1:20130) so its web UI can be iframed inside the
- * OmniRoute dashboard without CORS issues.
+ * Thin wrapper — all proxy logic lives in @/lib/services/reverseProxy.ts.
  *
  * Security:
  *   - Target URL is constructed from the service's registered port — never
  *     from user input — eliminating SSRF risk.
  *   - The route is classified LOCAL_ONLY in routeGuard.ts; the management
  *     policy blocks all non-loopback access before this handler runs.
- *   - Client cookies and Authorization headers are stripped before forwarding
- *     to prevent credential leakage between OmniRoute and the embedded service.
- *   - Upstream set-cookie, x-frame-options, content-security-policy, and
- *     cross-origin-* headers are stripped from responses so the iframe is not
- *     broken by the embedded service's own security policies.
- *   - HTML responses are rewritten (parse5) so path-absolute URLs work through
- *     the proxy prefix.
+ *   - See reverseProxy.ts for full security documentation.
  */
 
-import { getSupervisor } from "@/lib/services/registry";
-import { getOrCreateApiKey } from "@/lib/services/apiKey";
-import { rewriteHtml } from "@/lib/services/htmlRewriter";
-import { createErrorResponse } from "@/lib/api/errorResponse";
-import { sanitizeErrorMessage } from "@omniroute/open-sse/utils/error";
+import { proxyRequest } from "@/lib/services/reverseProxy";
 
 export const dynamic = "force-dynamic";
 
-/** Standard hop-by-hop headers that must never be forwarded. */
-const HOP_BY_HOP = new Set([
-  "connection",
-  "keep-alive",
-  "proxy-authenticate",
-  "proxy-authorization",
-  "te",
-  "trailers",
-  "transfer-encoding",
-  "upgrade",
-  "host",
-]);
-
-/**
- * Request headers stripped before forwarding to the embedded service.
- * Prevents OmniRoute session cookies and Authorization from leaking upstream.
- */
-const STRIPPED_REQUEST_HEADERS = new Set(["cookie", "authorization"]);
-
-/**
- * Response headers stripped before returning to the browser.
- *
- * - set-cookie: prevents the embedded service from setting cookies in the
- *   OmniRoute origin, which would conflict with session management.
- * - content-security-policy / content-security-policy-report-only: the
- *   embedded service's CSP is irrelevant inside the OmniRoute iframe.
- * - x-frame-options: would block the iframe entirely if set to DENY/SAMEORIGIN
- *   by the embedded service (OmniRoute controls framing via its own CSP).
- * - cross-origin-*: remove COOP/COEP/CORP that could break the framed page.
- */
-const STRIPPED_RESPONSE_HEADERS = new Set([
-  "set-cookie",
-  "content-security-policy",
-  "content-security-policy-report-only",
-  "x-frame-options",
-  "cross-origin-embedder-policy",
-  "cross-origin-opener-policy",
-  "cross-origin-resource-policy",
-]);
-
-const PROXY_TIMEOUT_MS = 30_000;
-
 type RouteParams = { name: string; path: string[] };
 
-async function proxyToService(request: Request, params: RouteParams): Promise<Response> {
+async function handleProxy(request: Request, params: RouteParams): Promise<Response> {
   const { name, path } = params;
-
-  const supervisor = getSupervisor(name);
-  if (!supervisor) {
-    return createErrorResponse({ status: 404, message: `Service '${name}' not found.` });
-  }
-
-  const { state, port } = supervisor.getStatus();
-  if (state !== "running") {
-    return createErrorResponse({
-      status: 503,
-      message: `Service '${name}' is not running (state: ${state}).`,
-    });
-  }
-
-  const incomingUrl = new URL(request.url);
-  const upstreamPath = path.length > 0 ? "/" + path.join("/") : "/";
-  const upstreamUrl = `http://127.0.0.1:${port}${upstreamPath}${incomingUrl.search}`;
-
-  // Build forwarded headers: strip hop-by-hop AND sensitive client headers.
-  const forwardHeaders = new Headers();
-  for (const [k, v] of request.headers.entries()) {
-    const lower = k.toLowerCase();
-    if (!HOP_BY_HOP.has(lower) && !STRIPPED_REQUEST_HEADERS.has(lower)) {
-      forwardHeaders.set(k, v);
-    }
-  }
-  forwardHeaders.set("host", `127.0.0.1:${port}`);
-
-  // Inject the embedded service's own API key so upstream can authenticate.
-  const apiKey = await getOrCreateApiKey(name);
-  forwardHeaders.set("authorization", `Bearer ${apiKey}`);
-
-  const hasBody = request.method !== "GET" && request.method !== "HEAD";
-
-  try {
-    const upstream = await fetch(upstreamUrl, {
-      method: request.method,
-      headers: forwardHeaders,
-      body: hasBody ? request.body : undefined,
-      // @ts-expect-error -- duplex is required by the Fetch spec for streaming
-      // request bodies but is not yet in the TS DOM lib (Node.js 18+ supports it).
-      duplex: hasBody ? "half" : undefined,
-      signal: AbortSignal.timeout(PROXY_TIMEOUT_MS),
-    });
-
-    // Build response headers: strip hop-by-hop and security-conflicting headers.
-    const responseHeaders = new Headers();
-    for (const [k, v] of upstream.headers.entries()) {
-      const lower = k.toLowerCase();
-      if (!HOP_BY_HOP.has(lower) && !STRIPPED_RESPONSE_HEADERS.has(lower)) {
-        responseHeaders.set(k, v);
-      }
-    }
-    // Prevent Next.js from caching the proxied response.
-    responseHeaders.set("cache-control", "no-store");
-
-    const contentType = upstream.headers.get("content-type") ?? "";
-
-    // HTML responses: buffer, rewrite links, return as string.
-    if (contentType.startsWith("text/html")) {
-      const html = await upstream.text();
-      const publicPrefix = `/dashboard/providers/services/${name}/embed`;
-      const rewritten = rewriteHtml(html, publicPrefix);
-      return new Response(rewritten, {
-        status: upstream.status,
-        headers: responseHeaders,
-      });
-    }
-
-    // All other content types: stream through unchanged.
-    return new Response(upstream.body, {
-      status: upstream.status,
-      headers: responseHeaders,
-    });
-  } catch (err) {
-    const msg = sanitizeErrorMessage(err instanceof Error ? err.message : String(err));
-    return createErrorResponse({ status: 502, message: `Proxy error: ${msg}` });
-  }
+  return proxyRequest(request, path, {
+    name,
+    publicPrefix: `/dashboard/providers/services/${name}/embed`,
+    htmlRewrite: true,
+  });
 }
 
 export async function GET(
   request: Request,
   { params }: { params: Promise<RouteParams> }
 ): Promise<Response> {
-  return proxyToService(request, await params);
+  return handleProxy(request, await params);
 }
 
 export async function POST(
   request: Request,
   { params }: { params: Promise<RouteParams> }
 ): Promise<Response> {
-  return proxyToService(request, await params);
+  return handleProxy(request, await params);
 }
 
 export async function PUT(
   request: Request,
   { params }: { params: Promise<RouteParams> }
 ): Promise<Response> {
-  return proxyToService(request, await params);
+  return handleProxy(request, await params);
 }
 
 export async function PATCH(
   request: Request,
   { params }: { params: Promise<RouteParams> }
 ): Promise<Response> {
-  return proxyToService(request, await params);
+  return handleProxy(request, await params);
 }
 
 export async function DELETE(
   request: Request,
   { params }: { params: Promise<RouteParams> }
 ): Promise<Response> {
-  return proxyToService(request, await params);
+  return handleProxy(request, await params);
 }
 
 export async function HEAD(
   request: Request,
   { params }: { params: Promise<RouteParams> }
 ): Promise<Response> {
-  return proxyToService(request, await params);
+  return handleProxy(request, await params);
 }
 
 export async function OPTIONS(
   request: Request,
   { params }: { params: Promise<RouteParams> }
 ): Promise<Response> {
-  return proxyToService(request, await params);
+  return handleProxy(request, await params);
 }
